Move plist_utils_test.dart to plist_parser_test.dart
